Routine commit - Optimize PackageFlow, which always refers to the top application - Optimize ScopedFlow - Destroy ShizukuAutomatorService if there is an error occurred when connecting - Deal with suspension state of a task - And more

+++ tasker-engine/src/main/java/top/xjunz/tasker/engine/runtime/Snapshot.kt
@@ -9,7 +9,7 @@ package top.xjunz.tasker.engine.runtime
  */
 class Snapshot {
 
-    val registry = mutableMapOf<Int, Any>()
+    val registry = mutableMapOf<Long, Any>()
 
     fun clear() {
         registry.clear()
